requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
PyPDF2==3.0.1
python-magic==0.4.27
mammoth==1.7.0
scipy==1.15.2
SQLAlchemy==2.0.29
Werkzeug==3.1.3
psycopg2-binarypsycopg2-binary 
